+++ packages/orq-rc/RELEASES.md
@@ -2309,3 +2309,13 @@ Based on:
 - [typescript v3.9.0-rc.0] packages/orq-rc
 ### Releases
 - [NPM v3.9.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/3.9.0-rc.0 - packages/orq-rc
+
+## 2025-06-23 12:23:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.568.0 (2.632.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.9.0-rc.1] packages/orq-rc
+### Releases
+- [NPM v3.9.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/3.9.0-rc.1 - packages/orq-rc